Refactor :     1. Remove evnsq/exp.h, evnsq/evpp_export.h and evmc/exp.h     2. Rename evpp/log_config.h to evpp/logging.h
