--- apps/evmc/exp.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#ifdef _WIN32
-#pragma comment(lib, "evmc_static.lib")
-#endif
-
