+++ apps/evmc/extract_vbucket_conf.h
@@ -1,6 +1,7 @@
-#ifndef EXTRACT_VBUCKET_CONF_H_
-#define EXTRACT_VBUCKET_CONF_H_
+#pragma once
+
 #include "evpp/httpc/request.h"
+
 namespace evmc {
 enum {
     READ_VBUCKET_CONF_FAILED = -1,
@@ -18,4 +19,3 @@ private:
     static int GetRemoteVbucketConf(const std::string& conf_addr, std::string& context);
 };
 }
-#endif
