--- apps/evnsq/exp.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "evnsq_export.h"
-#include "consumer.h"
-
-#ifdef H_OS_WINDOWS
-#pragma comment(lib, "evnsq_static.lib")
-#endif
-
