+++ apps/evnsq/test/producer_with_auth/main.cc
@@ -1,4 +1,3 @@
-#include <evnsq/exp.h>
 #include <evnsq/consumer.h>
 #include <evnsq/producer.h>
 #include <evpp/event_loop.h>
